inner polygon works for 2hex, but switching to stroke way
We'll make an svg shape for the tile and add a stroke. Too hard and slow to make the inner polygon and position it. Eventhough, in LibreOffice, the clip paths are not applied to the svgs, so the svgs are "too big" for their hexes (the stroke leaks past the radius)