Choose: Polygon stroke clippath, not polygon inside polygon... for now
Eventually, it would be nice if the shapes that came out in LibreOffice could be reassembled easily into a different hexgrid. Right now, it seems like the clip paths are not applied, and the stroke goes outside the desired radius of each hexagon, causing considerable overlap.

However, it displays fine in our generated pdf, so we're going with it!