Strange bug: number converted to string causes outline chaos
Create a new map, lay down a tile at some rotation, and the outlines are messed up. Add width to the map, and it fixes itself. So strange! All along from buildupJSONFileMap through addPiece, the rotation is a number. Gets applied as pieceRotation, still a number. But when you log pieceRotation inside TopOutlineInterlockHex, it is a string sometimes and sometimes a number!